backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.6
